refactor: move middleware to pkg
So that projects that use OpenFGA as a library can use our middleware
